Move function to avoid circular import
